requirements.txt
streamlit
PyMuPDF
sentence-transformers
faiss-cpu
numpy
google-generativeai
python-dotenv
requests
langchain-text-splitters
langchain-community
langchain-huggingface